requirements.txt
fastapi
uvicorn
sqlalchemy
bcrypt
python-jose[cryptography]
pydantic
pydantic-settings
python-dotenv
pytest
httpx
python-multipart
